refactor gcm cipher factory
